Remove "unix:" from the DISPLAY name.
This used to work (on Arch) but apparently erroneously.
It seems that a value like "unix:0.0" is for specific Unix variants
(SVR3 and SVR4) that use alternative local connection mechanisms;
whereas we are using UNIX Domain sockets for which a leading
colon (empty hostname) is the correct way to make local X connections.